Revert "Fix/118/static page bailout (#119)"
This reverts commit e0288486c68f416ae2e48f26b25a3a7a564ac98a.
