--- src/app/(default-layout)/layout.tsx
@@ -1,6 +1,0 @@
-import { PropsWithChildren } from "react";
-import { DefaultLayout } from "@/components/Layout";
-
-export default function Layout({ children }: PropsWithChildren) {
-  return <DefaultLayout>{children}</DefaultLayout>;
-}
