+++ src/app/(etc)/about/page.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import Image from "next/image";
 import Link from "next/link";
 import Logo from "@/assets/img/logo.svg";
