+++ src/app/(etc)/privacy/page.tsx
