--- src/app/(roadmap-layout)/layout.tsx
@@ -1,6 +1,0 @@
-import { PropsWithChildren } from "react";
-import { RoadmapLayout } from "@/components/Layout";
-
-export default function Layout({ children }: PropsWithChildren) {
-  return <RoadmapLayout>{children}</RoadmapLayout>;
-}
