+++ src/app/admin/layout.tsx
@@ -13,7 +13,7 @@ const MENU = [
   },
 ];
 
-export default function AdminLayout({ children }: PropsWithChildren) {
+export default function MypageLayout({ children }: PropsWithChildren) {
   return (
     <div className="min-h-[400px]">
       <header className="pt-4 md:pt-0">
