+++ src/app/layout.tsx
@@ -1,8 +1,12 @@
+import { Suspense } from "react";
 import { Metadata } from "next";
 import { Roboto_Serif } from "next/font/google";
-import { Toaster } from "sonner";
+import FilterProvider from "@/components/FilterProvider";
 import Footer from "@/components/Footer";
+import Header from "@/components/Header";
+import Inner from "@/components/Inner";
 import { ThemeProvider } from "@/components/ThemeProvider";
+import { Toaster } from "@/components/ui/sonner";
 import "./globals.css";
 
 const robotoSerif = Roboto_Serif({
@@ -60,16 +64,23 @@ export default function RootLayout({
         ></script>
       </head>
       <body className="group antialiased">
-        <ThemeProvider
-          attribute="class"
-          defaultTheme="system"
-          enableSystem
-          disableTransitionOnChange
-        >
-          {children}
-          <Footer />
-          <Toaster />
-        </ThemeProvider>
+        <Suspense>
+          <FilterProvider basePath="/">
+            <ThemeProvider
+              attribute="class"
+              defaultTheme="system"
+              enableSystem
+              disableTransitionOnChange
+            >
+              <Header />
+              <Inner className="pb-10">
+                <main>{children}</main>
+              </Inner>
+              <Footer />
+              <Toaster />
+            </ThemeProvider>
+          </FilterProvider>
+        </Suspense>
       </body>
     </html>
   );
